[GM-2350] Function in Function Check
- moving the function in function check to the ParamsTypeCheck
- updating the unit tests moving it to the new TestParamsTypeCheck
testing class
- removing the old checking from the FunctionCallCheck
